Clean up file and folder structure
