+++ lib/asteroid.rb
