+++ lib/table.rb
