+++ test/table_test.rb
@@ -1,12 +1,10 @@
 require 'minitest/autorun'
-require 'minitest/pride'
-require 'pry'
-require_relative 'asteroid'
+require_relative '../lib/table'
+require_relative '../lib/asteroid'
 
-class AsteroidTest < Minitest::Test
+class TableTest < Minitest::Test
 
   def setup
-
     @asteroid_data = [
       {:links=>{:self=>"http://www.neowsapp.com/rest/v1/neo/3561029?api_key=yZaX1lTPU14xNTXRHjHUnjp6pHbI8JhRgJ2njwRQ"},
         :id=>"3561029",
@@ -49,51 +47,49 @@ class AsteroidTest < Minitest::Test
           :orbiting_body=>"Earth"}],
         :is_sentry_object=>false}
       ]
-    
-    @asteroid_details =
-      {:links=>{:self=>"http://www.neowsapp.com/rest/v1/neo/3561029?api_key=yZaX1lTPU14xNTXRHjHUnjp6pHbI8JhRgJ2njwRQ"},
-        :id=>"3561029",
-        :neo_reference_id=>"3561029",
-        :name=>"(2011 GE3)",
-        :nasa_jpl_url=>"http://ssd.jpl.nasa.gov/sbdb.cgi?sstr=3561029",
-        :absolute_magnitude_h=>26.0,
-        :estimated_diameter=>
-         {:kilometers=>{:estimated_diameter_min=>0.0167708462, :estimated_diameter_max=>0.0375007522},
-          :meters=>{:estimated_diameter_min=>16.7708462163, :estimated_diameter_max=>37.5007521798},
-          :miles=>{:estimated_diameter_min=>0.0104209175, :estimated_diameter_max=>0.0233018799},
-          :feet=>{:estimated_diameter_min=>55.0224631002, :estimated_diameter_max=>123.0339677816}},
-        :is_potentially_hazardous_asteroid=>false,
-        :close_approach_data=>
-         [{:close_approach_date=>"2019-03-30",
-           :close_approach_date_full=>"2019-Mar-30 05:04",
-           :epoch_date_close_approach=>1553922240000,
-           :relative_velocity=>{:kilometers_per_second=>"18.6491093989", :kilometers_per_hour=>"67136.7938360274", :miles_per_hour=>"41716.1928143916"},
-           :miss_distance=>{:astronomical=>"0.382360762", :lunar=>"148.738336418", :kilometers=>"57200355.56677694", :miles=>"35542652.814484172"},
-           :orbiting_body=>"Earth"}],
-        :is_sentry_object=>false}
+
+      @asteroid_list = Asteroid.list(@asteroid_data)
+      @column_labels = { name: "Name", diameter: "Diameter", miss_distance: "Missed The Earth By:" }
   end
+  
   def test_it_exists
-    asteroid = Asteroid.new(@asteroid_details)
-    assert_instance_of Asteroid, asteroid
+    table = Table.new({},[])
+    assert_instance_of Table, table
   end
 
-  def test_it_has_details
-    asteroid = Asteroid.new(@asteroid_details)
-    assert_equal "(2011 GE3)", asteroid.name 
-    assert_equal 123, asteroid.diameter
-    assert_equal 35542652, asteroid.miss_distance
+  def test_it_has_attributes
+    table = Table.new(@column_labels, @asteroid_list)
+    assert_equal @column_labels, table.column_labels
+    assert_equal @asteroid_list, table.row_data
   end
-
-  def test_it_can_create_an_array_of_asteroids
-    assert Asteroid.all_asteroids(@asteroid_data).all? {|asteroid| asteroid.class == Asteroid}
-    assert_equal 2, Asteroid.all_asteroids(@asteroid_data).size
+  
+  def test_it_can_create_column_data
+    table = Table.new(@column_labels, @asteroid_list)
+    result = {:name=>{:label=>"Name", :width=>11}, :diameter=>{:label=>"Diameter", :width=>8}, :miss_distance=>{:label=>"Missed The Earth By:", :width=>20}}
+    assert_equal result, table.column_data
   end
-
-  def test_it_can_find_the_number_of_asteroids
-    assert_equal 2, Asteroid.count(@asteroid_data)
+  
+  def test_it_can_create_a_header
+    table = Table.new(@column_labels, @asteroid_list)
+    result =  "| Name        | Diameter | Missed The Earth By: |\n"
+    assert_equal result, table.header
   end
-
-  def test_it_can_find_the_largest
-    assert_equal 147, Asteroid.largest(@asteroid_data).diameter
+  
+  def test_it_can_create_a_divider
+    table = Table.new(@column_labels, @asteroid_list)
+    result = "+-------------+----------+----------------------+\n"
+    assert_equal result, table.divider
+  end
+  
+  def test_it_can_create_rows
+    table = Table.new(@column_labels, @asteroid_list)
+    result = ["| (2011 GE3)  | 123 ft   | 35542652 miles       |\n", "| (2011 GK44) | 147 ft   | 10701438 miles       |\n"]
+    assert_equal result, table.create_rows
+  end
+  
+  def test_it_can_display_table
+    table = Table.new(@column_labels, @asteroid_list)
+    result = "+-------------+----------+----------------------+\n| Name        | Diameter | Missed The Earth By: |\n| (2011 GE3)  | 123 ft   | 35542652 miles       |\n| (2011 GK44) | 147 ft   | 10701438 miles       |\n+-------------+----------+----------------------+\n"
+    assert_equal result, table.display_table
   end
 end
